fix(fpo): accept bfloat16 values from the policy
Pi0Policy builds its value head in bfloat16, so its values arrive in
bfloat16. numpy has no bfloat16 type, and FPO turns values into numpy arrays in
two places. The first learn against the real pi05_libero policy stopped at the
first of them:

  fpo.py _refresh_epoch_value_targets
    value_batch.detach().cpu().numpy()
  TypeError: Got unsupported ScalarType BFloat16

The second is GAEBuffer.compute_advantages_and_returns, which runs right
after it and does the same with policy.get_value().

Both now cast to float32 before leaving torch. The conversion happens at the
numpy boundary rather than in Pi0Policy, so any policy with a bfloat16 head
works, and the fix can be tested without openpi, which CI does not install.

The E11 harness check that found this ran on the cluster with batch_size 8,
past the out-of-memory failure batch_size 32 had hit. The test's tree
observation policy now returns bfloat16 values, as Pi0Policy's does. Without
the fix, four of its five cases fail on the line above; with it, all pass.

+++ src/plugrl_server/buffer/rollout_buffer.py
@@ -325,8 +325,9 @@ class GAEBuffer(RolloutBuffer):
                     next_observations[i : i + batch_size], aggregate_method="concat"
                 )
                 with torch.inference_mode():
+                    # A value head may run in bfloat16, which numpy cannot hold.
                     batch_values = (
-                        policy.get_value(batch_obs).cpu().numpy().reshape(-1, 1)
+                        policy.get_value(batch_obs).float().cpu().numpy().reshape(-1, 1)
                     )
                 self.last_values[next_ids[i : i + batch_size]] = batch_values
